chatデザイン追加
